refactor: move serial logic to its own module
